requirements.txt
mido
pyserial

